refactor(status): extract printStatus into status-utils, use shared icons
Move the status display logic from cli.ts into status-utils.ts where
the supporting utilities already live. Use icons from ansi.ts instead
of inline Unicode escapes.

cli.ts shrinks from 698 to 585 lines.
