+++ src/status-utils.ts
@@ -1,7 +1,10 @@
 import { existsSync, readFileSync } from "node:fs";
-import { join } from "node:path";
-import { getLogPath } from "./config.ts";
+import { join, basename } from "node:path";
+import prettyMs from "pretty-ms";
+import { getLogPath, getPidPath, readConfig, validateResolvedConfig } from "./config.ts";
 import { debug } from "./debug.ts";
+import { resolveWorkspaceConfig } from "./frontmatter.ts";
+import { icons } from "./ansi.ts";
 import type { LogEntry, Outcome } from "./types.ts";
 
 export type WorkspaceHealth = {
@@ -76,4 +79,148 @@ export function getLastOutcome(wsPath: string): { outcome: Outcome; ts: string }
     }
   }
   return null;
+}
+
+// --- Status display ---
+
+function readPid(): number | null {
+  try {
+    const raw = readFileSync(getPidPath(), "utf-8").trim();
+    const pid = Number(raw);
+    if (Number.isNaN(pid)) return null;
+    return pid;
+  } catch {
+    return null;
+  }
+}
+
+function isProcessAlive(pid: number): boolean {
+  try {
+    process.kill(pid, 0);
+    return true;
+  } catch (err: any) {
+    if (err?.code === "EPERM") return true;
+    return false;
+  }
+}
+
+function formatElapsed(isoDate: string): string {
+  const t = new Date(isoDate).getTime();
+  if (Number.isNaN(t)) return "invalid timestamp";
+  const diff = Date.now() - t;
+  return diff > 0 ? `${prettyMs(diff, { compact: true })} ago` : "just now";
+}
+
+/**
+ * Print full `murmur status` output to stdout.
+ */
+export function printStatus() {
+  const pid = readPid();
+  const alive = pid ? isProcessAlive(pid) : false;
+
+  if (alive) {
+    console.log(`Daemon: running (PID ${pid})`);
+  } else {
+    console.log("Daemon: stopped");
+  }
+
+  const config = readConfig();
+  if (config.workspaces.length === 0) {
+    console.log(`\nNo workspaces configured. Run: murmur init <path>`);
+    return;
+  }
+
+  let validCount = 0;
+  let issueCount = 0;
+
+  const rows = config.workspaces.map((ws) => {
+    try {
+      const health = checkWorkspaceHealth(ws.path);
+      const resolved = resolveWorkspaceConfig(ws);
+      const configError = validateResolvedConfig(resolved);
+
+      let issue: string | undefined;
+      if (!health.pathExists) {
+        issue = "path does not exist";
+      } else if (!health.heartbeatExists) {
+        issue = "HEARTBEAT.md missing";
+      } else if (configError) {
+        issue = configError;
+      }
+
+      if (issue) {
+        issueCount++;
+      } else {
+        validCount++;
+      }
+
+      const statusIcon = issue ? icons.fail : icons.ok;
+      const name = resolved.name ?? basename(ws.path);
+      const schedule = resolved.interval
+        ? `every ${resolved.interval}`
+        : resolved.cron
+          ? `cron ${resolved.cron}`
+          : "(none)";
+
+      let statusMsg: string;
+      if (issue) {
+        statusMsg = issue;
+      } else if (!ws.lastRun) {
+        statusMsg = "never run";
+      } else {
+        const ago = formatElapsed(ws.lastRun);
+        const lastOutcome = getLastOutcome(ws.path);
+        const outcomeTag = lastOutcome ? ` (${lastOutcome.outcome})` : "";
+        statusMsg = `last: ${ago}${outcomeTag}`;
+      }
+
+      return { statusIcon, name, schedule, statusMsg, path: ws.path };
+    } catch {
+      issueCount++;
+      return {
+        statusIcon: icons.fail,
+        name: basename(ws.path),
+        schedule: "(unknown)",
+        statusMsg: "config error",
+        path: ws.path,
+      };
+    }
+  });
+
+  const nameW = Math.max(...rows.map((r) => r.name.length));
+  const schedW = Math.max(...rows.map((r) => r.schedule.length));
+  const statusW = Math.max(...rows.map((r) => r.statusMsg.length));
+
+  const summary =
+    issueCount > 0
+      ? `${validCount} valid, ${issueCount} ${issueCount === 1 ? "issue" : "issues"}`
+      : `${validCount}`;
+  console.log(`\nWorkspaces (${summary}):`);
+
+  for (const r of rows) {
+    const nameCol = r.name.padEnd(nameW);
+    const schedCol = r.schedule.padEnd(schedW);
+    const statusCol = r.statusMsg.padEnd(statusW);
+    console.log(`  ${r.statusIcon} ${nameCol}  ${schedCol}  ${statusCol}  ${r.path}`);
+  }
+
+  // Recent issues from heartbeat log
+  const nameByPath = new Map(rows.map((r) => [r.path, r.name]));
+  const errors = readRecentErrors(5);
+  if (errors.length > 0) {
+    console.log(`\nRecent issues (last 24h):`);
+    for (const entry of errors) {
+      const elapsed = Date.now() - new Date(entry.ts).getTime();
+      const time =
+        Number.isNaN(elapsed) || elapsed < 0
+          ? "just now"
+          : `${prettyMs(elapsed, { compact: true })} ago`;
+      const msg =
+        entry.outcome === "error"
+          ? (entry.error ?? "unknown error")
+          : (entry.summary ?? "needs attention");
+      const wsName = nameByPath.get(entry.workspace) ?? basename(entry.workspace);
+      console.log(`  ${icons.bullet} ${time} ${wsName}: ${msg}`);
+    }
+  }
 }
